==> scripts/00-install_dependencies.qmd <==
---
title: "Install the dependencies required for this project"
format: html
author: "Sebastian Rodriguez"
---
```{r}
#### Preamble ####
# Purpose: Install the dependencies required for this project
# Prerequisites: You must also have the R language installed
# Author: Sebastian Rodriguez
# Email: me@srod.ca
# Date: 3 February 2023
# GitHub: https://github.com/seb646/toronto-paramedic-responses
```

```{r}
# Install required packages (you only need to do this once)
install.packages("lubridate")
install.packages("tidyverse")
install.packages("dplyr")
install.packages("knitr")
install.packages("janitor")
install.packages("scales")
install.packages("RColorBrewer")
install.packages("ggplot2")
install.packages("kableExtra")
install.packages("here")
install.packages("readr")
install.packages("foreign")
```
